Pass series and publishers as query params.
