--- src/main/java/de/wasenweg/alfred/publisher/PublisherRepository.java
@@ -1,7 +1,0 @@
-package de.wasenweg.alfred.publisher;
-
-import org.springframework.data.mongodb.repository.MongoRepository;
-
-public interface PublisherRepository extends MongoRepository<Publisher, String>, PublisherQueryRepository {
-
-}
