+++ src/test/java/de/wasenweg/alfred/integration/VolumesIntegrationTest.java
@@ -2,14 +2,12 @@ package de.wasenweg.alfred.integration;
 
 import de.wasenweg.alfred.EnableEmbeddedMongo;
 import de.wasenweg.alfred.TestUtil;
-import de.wasenweg.alfred.comics.Comic;
 import de.wasenweg.alfred.comics.ComicRepository;
 import de.wasenweg.alfred.fixtures.ComicFixtures;
 import de.wasenweg.alfred.fixtures.ProgressFixtures;
 import de.wasenweg.alfred.fixtures.VolumeFixtures;
 import de.wasenweg.alfred.progress.Progress;
 import de.wasenweg.alfred.progress.ProgressRepository;
-import de.wasenweg.alfred.scanner.ScannerIssue;
 import lombok.RequiredArgsConstructor;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
@@ -65,36 +63,10 @@ public class VolumesIntegrationTest {
   }
 
   @Test
-  public void findAllPublishers() throws Exception {
-    this.mockMvc.perform(get("/api/publishers"))
-        .andExpect(status().isOk())
-        .andExpect(content().contentType(MediaTypes.HAL_JSON_VALUE))
-        .andExpect(jsonPath("$._embedded.publishers.length()").value(1))
-        .andExpect(jsonPath("$._embedded.publishers[0].name")
-            .value(ComicFixtures.COMIC_V1_1.getPublisher()))
-        .andExpect(jsonPath("$._embedded.publishers[0].seriesCount").value(1))
-        .andExpect(jsonPath("$._embedded.publishers[0].series.length()").value(1))
-        .andExpect(jsonPath("$._embedded.publishers[0].series[0].volumesCount").value(3));
-  }
-
-  @Test
-  public void findAllSeries() throws Exception {
-    this.mockMvc.perform(get("/api/publishers/" + ComicFixtures.COMIC_V1_1.getPublisher() + "/series"))
-        .andExpect(status().isOk())
-        .andExpect(content().contentType(MediaTypes.HAL_JSON_VALUE))
-        .andExpect(jsonPath("$._embedded.series.length()").value(1))
-        .andExpect(jsonPath("$._embedded.series[0].name")
-            .value(ComicFixtures.COMIC_V1_1.getSeries()))
-        .andExpect(jsonPath("$._embedded.series[0].publisher")
-            .value(ComicFixtures.COMIC_V1_1.getPublisher()))
-        .andExpect(jsonPath("$._embedded.series[0].volumesCount").value(3));
-  }
-
-  @Test
   public void findAllVolumes() throws Exception {
-    this.mockMvc.perform(get("/api/publishers/"
-            + ComicFixtures.COMIC_V1_1.getPublisher() + "/series/"
-            + ComicFixtures.COMIC_V1_1.getSeries() + "/volumes"))
+    this.mockMvc.perform(get("/api/volumes")
+        .param("publisher", ComicFixtures.COMIC_V1_1.getPublisher())
+        .param("series", ComicFixtures.COMIC_V1_1.getSeries()))
         .andExpect(status().isOk())
         .andExpect(content().contentType(MediaTypes.HAL_JSON_VALUE))
         .andExpect(jsonPath("$._embedded.volumes.length()").value(3))
@@ -131,53 +103,6 @@ public class VolumesIntegrationTest {
         .andExpect(jsonPath("$._embedded.volumes[2].issueCount").value(3))
         .andExpect(jsonPath("$._embedded.volumes[2].readCount").value(0))
         .andExpect(jsonPath("$._embedded.volumes[2].read").value(false));
-  }
-
-  @Test
-  public void findAllPublishersWithMoreThanOne() throws Exception {
-    // When there are two publishers, each of which have the exact same
-    // series name and volume name.
-    final Comic pivotal = this.comicRepository.findByPath("/a1.cbz").get();
-    pivotal.setPublisher("Publisher B");
-    this.comicRepository.save(pivotal);
-
-    this.mockMvc.perform(get("/api/publishers"))
-        .andExpect(status().isOk())
-        .andExpect(content().contentType(MediaTypes.HAL_JSON_VALUE))
-        .andExpect(jsonPath("$._embedded.publishers.length()").value(2))
-        .andExpect(jsonPath("$._embedded.publishers[0].name")
-            .value(ComicFixtures.COMIC_V1_1.getPublisher()))
-        .andExpect(jsonPath("$._embedded.publishers[0].seriesCount").value(1))
-        .andExpect(jsonPath("$._embedded.publishers[0].series.length()").value(1))
-        .andExpect(jsonPath("$._embedded.publishers[0].series[0].name")
-            .value(ComicFixtures.COMIC_V1_1.getSeries()))
-        .andExpect(jsonPath("$._embedded.publishers[0].series[0].volumesCount").value(3))
-        .andExpect(jsonPath("$._embedded.publishers[1].name")
-            .value("Publisher B"))
-        .andExpect(jsonPath("$._embedded.publishers[1].seriesCount").value(1))
-        .andExpect(jsonPath("$._embedded.publishers[1].series.length()").value(1))
-        .andExpect(jsonPath("$._embedded.publishers[1].series[0].volumesCount").value(1))
-        .andExpect(jsonPath("$._embedded.publishers[1].series[0].name")
-            .value(ComicFixtures.COMIC_V1_1.getSeries()));
-  }
-
-  @Test
-  public void findAllPublishersWithErrors() throws Exception {
-    final Comic pivotal = this.comicRepository.findByPath("/a1.cbz").get();
-    pivotal.setPublisher("Pub B");
-    pivotal.setErrors(Arrays.asList(
-        ScannerIssue.builder().severity(ScannerIssue.Severity.ERROR).message("Mock Error").build()));
-    this.comicRepository.save(pivotal);
-
-    this.mockMvc.perform(get("/api/publishers"))
-        .andExpect(status().isOk())
-        .andExpect(content().contentType(MediaTypes.HAL_JSON_VALUE))
-        .andExpect(jsonPath("$._embedded.publishers.length()").value(1))
-        .andExpect(jsonPath("$._embedded.publishers[0].name")
-            .value(ComicFixtures.COMIC_V1_1.getPublisher()))
-        .andExpect(jsonPath("$._embedded.publishers[0].seriesCount").value(1))
-        .andExpect(jsonPath("$._embedded.publishers[0].series.length()").value(1))
-        .andExpect(jsonPath("$._embedded.publishers[0].series[0].volumesCount").value(3));
   }
 
   @Test
