+++ ui/src/app/library/publishers/publishers.page.html
@@ -42,7 +42,7 @@
       </ion-item>
       <ion-item *ngFor="let serie of publisher.series" class="serie">
         <ion-button fill="solid" color="secondary"
-          [routerLink]="['/library/publishers/' + serie.publisher + '/series/' + serie.name + '/volumes']">
+          [routerLink]="['/library/publishers', serie.publisher, 'series', serie.name, 'volumes']">
           {{ serie.name }}
         </ion-button>
       </ion-item>
