+++ ui/src/app/library/series/series.page.html
@@ -40,7 +40,7 @@
     </ion-item>
     <ion-item *ngFor="let serie of series" class="serie">
       <ion-button fill="solid" color="secondary"
-        [routerLink]="['/library/publishers/' + serie.publisher + '/series/' + serie.name + '/volumes']">
+        [routerLink]="['/library/publishers', serie.publisher,  'series', serie.name, 'volumes']">
         {{ serie.name }}
       </ion-button>
     </ion-item>
